Update k8s config load-balancer-l2-interfaces
For local mode, get the interface name based on
management address and update k8s config
load-balancer-l2-interfaces with the interface
name.
For maas mode, retrieve the interface name from
MAAS for the space Management for any one of the
control nodes and update the k8s config
load-balancer-l2-interfaces.

Fixes: LP#2098438