Add host-only to microceph list-disks action
host-only param is introduced to list-disks action
in [1]. Use the flag in sunbeam while configuring
disks so that only disks on that host will be
retrieved.

The change is backward compatible as using the new
param on action with older charm-microceph returns
all the OSDs of all hosts just like earlier.

[1] https://github.com/canonical/charm-microceph/pull/147

Fixes: https://bugs.launchpad.net/snap-openstack/+bug/2098626