Check if compute role exists before launching VM
There is currently no check to see if a compute role is present
in the cluster before launching a VM, resulting in an error. This
change will fail the VM launch gracefully and present the user with
a descriptive error message if the compute role is not present. A
check to ensure the cluster is bootstrapped is also added.

LP#2100006