Reapply Hypervisor plan with optional integrations as target
When multiple nodes are joined with compute+storage role, the
DeployOpenStackHypervisorApplication step updates the machine-id's
with known juju machines at that point of time. However due to
parallel node joins, the machines might have been updated in juju
just before applying the terraform plan which makes discrepancy
in machine_id's in terraform vars resulted in removal of machines.

This step is introduced in join workflow to update any missing
integrations, especially ceph. So reapply the terraform plan with
target as integrations instead of the complete plan. It is safe to
update only optional integrations in this step since they are
dynamic in nature and applied when certain roles are tagged or
features are enabled.

Fixes: https://bugs.launchpad.net/snap-openstack/+bug/2103499